Refactor/#40/회원가입 및 정보수정 로직 통합
